+++ sdk/communication/azure-communication-identity/tests/test_communication_identity_client_async.py
@@ -139,3 +139,48 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
             await identity_client.delete_user(user)
 
         assert user.properties.get('id') is not None
+    
+    @CommunicationPreparer()
+    async def test_create_user_and_token_with_no_scopes(self, communication_connection_string):
+        identity_client = CommunicationIdentityClient.from_connection_string(communication_connection_string)
+
+        async with identity_client:
+            with pytest.raises(Exception) as ex:
+                user, token_response = await identity_client.create_user_and_token(scopes=None)
+
+    @CommunicationPreparer()
+    async def test_delete_user_with_no_user(self, communication_connection_string):
+        identity_client = CommunicationIdentityClient.from_connection_string(
+            communication_connection_string)
+        
+        async with identity_client:
+            with pytest.raises(Exception) as ex:
+                await identity_client.delete_user(user=None)
+
+    @CommunicationPreparer()
+    async def test_revoke_tokens_with_no_user(self, communication_connection_string):
+        identity_client = CommunicationIdentityClient.from_connection_string(
+            communication_connection_string)
+        
+        async with identity_client:
+            with pytest.raises(Exception) as ex:
+                await identity_client.revoke_tokens(user=None)
+    
+    @CommunicationPreparer()
+    async def test_get_token_with_no_user(self, communication_connection_string):
+        identity_client = CommunicationIdentityClient.from_connection_string(
+            communication_connection_string)
+        
+        async with identity_client:
+            with pytest.raises(Exception) as ex:
+                token_response = await identity_client.get_token(user=None, scopes=[CommunicationTokenScope.CHAT])
+    
+    @CommunicationPreparer()
+    async def test_get_token_with_no_scopes(self, communication_connection_string):
+        identity_client = CommunicationIdentityClient.from_connection_string(
+            communication_connection_string)
+
+        async with identity_client:
+            with pytest.raises(Exception) as ex:
+                user = await identity_client.create_user()
+                token_response = await identity_client.get_token(user, scopes=None)