+++ sdk/storage/azure-storage-blob/azure/storage/blob/_blob_client.py
@@ -3754,7 +3754,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             process_storage_error(error)
 
     @distributed_trace
-    def get_container_client(self): # pylint: disable=client-method-missing-kwargs
+    def _get_container_client(self): # pylint: disable=client-method-missing-kwargs
         # type: (...) -> ContainerClient
         """Get a client to interact with the blob's parent container.