+++ sdk/tables/azure-data-tables/azure/data/tables/__init__.py
@@ -6,6 +6,7 @@
 from azure.data.tables._models import TableServiceStats
 
 from ._entity import TableEntity, EntityProperty, EdmType
+from ._error import RequestTooLargeError
 from ._table_shared_access_signature import generate_table_sas, generate_account_sas
 from ._table_client import TableClient
 from ._table_service_client import TableServiceClient
@@ -22,7 +23,6 @@ from ._models import (
     LocationMode,
     ResourceTypes,
     AccountSasPermissions,
-    BatchTransactionResult,
     BatchErrorException,
 )
 from ._version import VERSION
@@ -53,7 +53,7 @@ __all__ = [
     "RetentionPolicy",
     "generate_table_sas",
     "SASProtocol",
-    "BatchTransactionResult",
     "TableBatchOperations",
     "BatchErrorException",
+    "RequestTooLargeError",
 ]