+++ sdk/tables/azure-data-tables/azure/data/tables/aio/_table_batch_async.py
@@ -8,6 +8,7 @@ import msrest
 
 from .._common_conversion import _is_cosmos_endpoint, _transform_patch_to_cosmos_post
 from .._models import UpdateMode
+from .._entity import TableEntity
 from .._serialize import (
     _get_match_headers,
     _add_entity_properties,
@@ -39,7 +40,6 @@ class TableBatchOperations(object):
         deserializer: msrest.Deserializer,
         config: AzureTableConfiguration,
         table_name: str,
-        table_client,  # type: TableClient,
         **kwargs: Dict[str, Any]
     ) -> None:
         self._client = client
@@ -47,21 +47,22 @@ class TableBatchOperations(object):
         self._deserialize = deserializer
         self._config = config
         self.table_name = table_name
-        self._table_client = table_client
 
         self._partition_key = kwargs.pop("partition_key", None)
         self._requests = []
         self._entities = []
 
+    def __len__(self):
+        return len(self._requests)
+
     def _verify_partition_key(
-        self, entity  # type: Union[Entity, dict]
+        self, entity  # type: Union[TableEntity, dict]
     ):
         # (...) -> None
         if self._partition_key is None:
             self._partition_key = entity["PartitionKey"]
-        elif "PartitionKey" in entity:
-            if entity["PartitionKey"] != self._partition_key:
-                raise ValueError("Partition Keys must all be the same")
+        elif entity["PartitionKey"] != self._partition_key:
+            raise ValueError("Partition Keys must all be the same")
 
     def create_entity(
         self,
@@ -86,13 +87,14 @@ class TableBatchOperations(object):
                 :caption: Creating and adding an entity to a Table
         """
         self._verify_partition_key(entity)
+        temp = entity.copy()
 
-        if "PartitionKey" in entity and "RowKey" in entity:
-            entity = _add_entity_properties(entity)
+        if "PartitionKey" in temp and "RowKey" in temp:
+            temp = _add_entity_properties(temp)
         else:
-            raise ValueError("PartitionKey and RowKey were not provided in entity")
-        self._batch_create_entity(table=self.table_name, entity=entity, **kwargs)
-        self._entities.append(entity)
+            raise ValueError("PartitionKey and/or RowKey were not provided in entity")
+        self._batch_create_entity(table=self.table_name, entity=temp, **kwargs)
+        self._entities.append(TableEntity(**entity.copy()))
 
     def _batch_create_entity(
         self,
@@ -213,6 +215,7 @@ class TableBatchOperations(object):
                 :caption: Creating and adding an entity to a Table
         """
         self._verify_partition_key(entity)
+        temp = entity.copy()
 
         if_match, _ = _get_match_headers(
             kwargs=dict(
@@ -224,16 +227,16 @@ class TableBatchOperations(object):
             match_param="match_condition",
         )
 
-        partition_key = entity["PartitionKey"]
-        row_key = entity["RowKey"]
-        entity = _add_entity_properties(entity)
+        partition_key = temp["PartitionKey"]
+        row_key = temp["RowKey"]
+        temp = _add_entity_properties(temp)
         if mode is UpdateMode.REPLACE:
             self._batch_update_entity(
                 table=self.table_name,
                 partition_key=partition_key,
                 row_key=row_key,
                 if_match=if_match or "*",
-                table_entity_properties=entity,
+                table_entity_properties=temp,
                 **kwargs
             )
         elif mode is UpdateMode.MERGE:
@@ -242,10 +245,10 @@ class TableBatchOperations(object):
                 partition_key=partition_key,
                 row_key=row_key,
                 if_match=if_match or "*",
-                table_entity_properties=entity,
+                table_entity_properties=temp,
                 **kwargs
             )
-        self._entities.append(entity)
+        self._entities.append(TableEntity(**entity.copy()))
 
     def _batch_update_entity(
         self,
@@ -515,9 +518,7 @@ class TableBatchOperations(object):
             if_match=if_match or "*",
             **kwargs
         )
-
-        temp_entity = {"PartitionKey": partition_key, "RowKey": row_key}
-        self._entities.append(_add_entity_properties(temp_entity))
+        self._entities.append(TableEntity(PartitionKey=partition_key, RowKey=row_key))
 
     def _batch_delete_entity(
         self,
@@ -637,17 +638,18 @@ class TableBatchOperations(object):
                 :caption: Creating and adding an entity to a Table
         """
         self._verify_partition_key(entity)
+        temp = entity.copy()
 
-        partition_key = entity["PartitionKey"]
-        row_key = entity["RowKey"]
-        entity = _add_entity_properties(entity)
+        partition_key = temp["PartitionKey"]
+        row_key = temp["RowKey"]
+        temp = _add_entity_properties(temp)
 
         if mode is UpdateMode.MERGE:
             self._batch_merge_entity(
                 table=self.table_name,
                 partition_key=partition_key,
                 row_key=row_key,
-                table_entity_properties=entity,
+                table_entity_properties=temp,
                 **kwargs
             )
         elif mode is UpdateMode.REPLACE:
@@ -655,19 +657,7 @@ class TableBatchOperations(object):
                 table=self.table_name,
                 partition_key=partition_key,
                 row_key=row_key,
-                table_entity_properties=entity,
+                table_entity_properties=temp,
                 **kwargs
             )
-        self._entities.append(entity)
-
-    async def __aenter__(self):
-        # type: (...) -> TableBatchOperations
-        return self
-
-    async def __aexit__(
-        self,
-        *args,  # type: Any
-        **kwargs  # type: Any
-    ):
-        # (...) -> None
-        await self._table_client._batch_send(*self._requests, **kwargs)
+        self._entities.append(TableEntity(**entity.copy()))